Fix seq compat validation in BindLineage
The actual problem was an utterly boneheaded bug in the Go code doing
the validation. Totally not CUE's fault, at all.

This restricts something we probably want to allow, at least optionally:
treating the addition of required fields with default values as
backwards compatible. But the more restrictive approach in this commit
can always be relaxed later, and no existing code will be broken.

Tutorial docs do still need to be updated.

Fixes #14