openapi: Total overhaul of codegen
Refactor codegen to no longer rely on eval and printing, taking
advantage of the ability to pass a cue.Value to the CUE standard library
package.

Also introduce support for a group flag, which allows translating only the
children of the root schema to OpenAPI.

And introduce a long-overdue test txtar-based test framework.