mod: Remove replace statement for cue-lang/cue
This was causing the main CLI not to be installable via `go install`.
It's rather ridiculous that this is how `go` works, but given that it
does, this is the less frictionful path in the interim while we await a
change in cue-lang/cue itself to incorporate the improved openapi
generation.