thema: Disable schema emptiness invariants
Attempting to migrate grafana/grafana to the new flattened lineage
format has revealed that there were a number of existing lineages that
contained no actual fields - only definitions. This violates the schema
non-emptiness invariant introduced during lineage flattening.

These lineages will be refactored, but to ease the transition, this
commit temporarily disables the schema non-emptiness invariant in thema
so that it can be handled gradually, rather than right away and up
front. This does not meaningfully degrade thema's guarantees, as we're
just returning to where it was pre-flattening.